Introducing generalization and re-engineering for Marker and Circle classes and their related List classes
