+++ GoogleMapsComponents/Maps/BaseListableEntity.cs
@@ -1,0 +1,106 @@
+﻿using Microsoft.JSInterop;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GoogleMapsComponents.Maps
+{
+    public abstract class BaseListableEntity<T> : IDisposable, IJsObjectRef
+        where T: BaseListableEntityOptions
+    {
+        protected readonly JsObjectRef _jsObjectRef;
+
+        public readonly Dictionary<string, List<MapEventListener>> EventListeners;
+
+        public Guid Guid => _jsObjectRef.Guid;
+
+        internal BaseListableEntity(JsObjectRef jsObjectRef)
+        {
+            _jsObjectRef = jsObjectRef;
+            EventListeners = new Dictionary<string, List<MapEventListener>>();
+        }
+
+        public void Dispose()
+        {
+            foreach (string key in EventListeners.Keys)
+            {
+                //Probably superfluous...
+                if (EventListeners[key] != null)
+                {
+                    foreach (MapEventListener eventListener in EventListeners[key])
+                    {
+                        eventListener.Dispose();
+                    }
+
+                    EventListeners[key].Clear();
+                }
+            }
+
+            EventListeners.Clear();
+            _jsObjectRef.Dispose();
+        }
+
+        public Task<Map> GetMap()
+        {
+            return _jsObjectRef.InvokeAsync<Map>(
+                "getMap");
+        }
+
+
+
+        /// <summary>
+        /// Renders the mao entity on the specified map or panorama. 
+        /// If map is set to null, the map entity will be removed.
+        /// </summary>
+        /// <param name="map"></param>
+        public async Task SetMap(Map map)
+        {
+            await _jsObjectRef.InvokeAsync(
+                   "setMap",
+                   map);
+
+            //_map = map;
+        }
+
+        public async Task<MapEventListener> AddListener(string eventName, Action handler)
+        {
+            JsObjectRef listenerRef = await _jsObjectRef.InvokeWithReturnedObjectRefAsync("addListener", eventName, handler);
+            MapEventListener eventListener = new MapEventListener(listenerRef);
+
+            if (!EventListeners.ContainsKey(eventName))
+            {
+                EventListeners.Add(eventName, new List<MapEventListener>());
+            }
+            EventListeners[eventName].Add(eventListener);
+
+            return eventListener;
+        }
+
+        public async Task<MapEventListener> AddListener<V>(string eventName, Action<V> handler)
+        {
+            JsObjectRef listenerRef = await _jsObjectRef.InvokeWithReturnedObjectRefAsync("addListener", eventName, handler);
+            MapEventListener eventListener = new MapEventListener(listenerRef);
+
+            if (!EventListeners.ContainsKey(eventName))
+            {
+                EventListeners.Add(eventName, new List<MapEventListener>());
+            }
+            EventListeners[eventName].Add(eventListener);
+
+            return eventListener;
+        }
+
+        public async Task ClearListeners(string eventName)
+        {
+            if (EventListeners.ContainsKey(eventName))
+            {
+                await _jsObjectRef.InvokeAsync("clearListeners", eventName);
+
+                //IMHO is better preserving the knowledge that Marker had some EventListeners attached to "eventName" in the past
+                //so, instead to clear the list and remove the key from dictionary, I prefer to leave the key with an empty list
+                EventListeners[eventName].Clear();
+            }
+        }
+    }
+}
