+++ GoogleMapsComponents/Maps/Extension/MarkerList.cs
@@ -16,11 +16,15 @@ namespace GoogleMapsComponents.Maps.Extension
     ///
     /// All properties should be called With a Dictionary<string, {property type}> indicating for each Marker(related to that key) the corresponding related property value
     /// </summary>
-    public class MarkerList : IDisposable
+    public class MarkerList : BaseListableEntityList<Marker, MarkerOptions>
     {
-        private readonly JsObjectRef _jsObjectRef;
-
-        public readonly Dictionary<string, Marker> Markers;
+        public Dictionary<string, Marker> Markers
+        {
+            get
+            {
+                return base.BaseListableEntities;
+            }
+        }
 
         /// <summary>
         /// Create markers list
@@ -51,18 +55,8 @@ namespace GoogleMapsComponents.Maps.Extension
         }
 
         private MarkerList(JsObjectRef jsObjectRef, Dictionary<string, Marker> markers)
+            :base(jsObjectRef, markers)
         {
-            _jsObjectRef = jsObjectRef;
-            Markers = markers;
-        }
-
-        public void Dispose()
-        {
-            if (Markers.Count > 0)
-            {
-                _jsObjectRef.DisposeMultipleAsync(Markers.Select(e => e.Value.Guid).ToList());
-                Markers.Clear();
-            }
         }
 
         /// <summary>
@@ -72,108 +66,8 @@ namespace GoogleMapsComponents.Maps.Extension
         /// <returns></returns>
         public async Task AddMultipleAsync(Dictionary<string, MarkerOptions> opts)
         {
-            if (opts.Count > 0)
-            {
-                Dictionary<string, JsObjectRef> jsObjectRefs = await _jsObjectRef.AddMultipleAsync(
-                    "google.maps.Marker",
-                    opts.ToDictionary(e => e.Key, e => (object)e.Value));
-                Dictionary<string, Marker> objs = jsObjectRefs.ToDictionary(e => e.Key, e => new Marker(e.Value));
-
-                //Someone can try to create element yet inside Markers... really not the best approach... but manage it
-                List<string> alreadyCreated = Markers.Keys.Intersect(objs.Select(e => e.Key)).ToList();
-                await RemoveMultipleAsync(alreadyCreated);
-
-                //Now we can add all required object as NEW object
-                foreach (string key in objs.Keys)
-                {
-                    Markers.Add(key, objs[key]);
-                }
-            }
-        }
-
-        /// <summary>
-        /// only Marker having keys matching with existent keys will be removed
-        /// </summary>
-        /// <param name="filterKeys"></param>
-        /// <returns></returns>
-        public async Task RemoveMultipleAsync(List<string> filterKeys = null)
-        {
-            if ((filterKeys != null) && (filterKeys.Count > 0))
-            {
-                List<string> foundKeys = Markers.Keys.Intersect(filterKeys).ToList();
-                if (foundKeys.Count > 0)
-                {
-                    List<Guid> foundGuids = Markers.Where(e => foundKeys.Contains(e.Key)).Select(e => e.Value.Guid).ToList();
-                    await _jsObjectRef.DisposeMultipleAsync(foundGuids);
-
-                    foreach (string key in foundKeys)
-                    {
-                        //Marker object needs to dispose call due to previous DisposeMultipleAsync call
-                        //Probably superfluous, but Garbage Collector may appreciate it... 
-                        Markers[key] = null;
-                        Markers.Remove(key);
-                    }
-                }
-            }
-        }
-
-        public async Task RemoveMultipleAsync(List<Guid> guids)
-        {
-            if (guids.Count > 0)
-            {
-                List<string> foundKeys = Markers.Where(e => guids.Contains(e.Value.Guid)).Select(e => e.Key).ToList();
-                if (foundKeys.Count > 0)
-                {
-                    List<Guid> foundGuids = Markers.Values.Where(e => guids.Contains(e.Guid)).Select(e => e.Guid).ToList();
-                    await _jsObjectRef.DisposeMultipleAsync(foundGuids);
-
-                    foreach (string key in foundKeys)
-                    {
-                        //Marker object needs to dispose call due to previous DisposeMultipleAsync call
-                        //Probably superfluous, but Garbage Collector may appreciate it... 
-                        Markers[key] = null;
-                        Markers.Remove(key);
-                    }
-                }
-            }
-        }
-
-        //Find the eventual match between required keys (if any) and yet stored markers key (if any)
-        //If filterKeys is null or empty all keys are returned
-        //Otherwise only eventually yet stored marker keys that matches with filterKeys
-        private List<string> ComputeMathingKeys(List<string> filterKeys = null)
-        {
-            List<string> matchingKeys;
-
-            if ((filterKeys == null) || (!filterKeys.Any()))
-            {
-                matchingKeys = Markers.Keys.ToList();
-            }
-            else
-            {
-                matchingKeys = Markers.Keys.Where(e => filterKeys.Contains(e)).ToList();
-            }
-
-            return matchingKeys;
-        }
-
-        //Creates mapping between matching keys and markers Guid
-        private Dictionary<Guid, string> ComputeInternalMapping(List<string> matchingKeys)
-        {
-            return Markers.Where(e => matchingKeys.Contains(e.Key)).ToDictionary(e => Markers[e.Key].Guid, e => e.Key);
-        }
-
-        //Creates mapping between markers Guid and empty array of parameters (getter has no parameter)
-        private Dictionary<Guid, object> ComputeDictArgs(List<string> matchingKeys)
-        {
-            return Markers.Where(e => matchingKeys.Contains(e.Key)).ToDictionary(e => e.Value.Guid, e => (object)(new object[] { }));
-        }
-
-        //Create an empty result of the correct type in case of no matching keys
-        private Task<Dictionary<string, T>> ComputeEmptyResult<T>()
-        {
-            return Task<Dictionary<string, T>>.Factory.StartNew(() => { return new Dictionary<string, T>(); });
-        }
+            await base.AddMultipleAsync(opts, "google.maps.Marker");           
+        }        
 
         public Task<Dictionary<string, Animation>> GetAnimations(List<string> filterKeys = null)
         {
@@ -230,26 +124,7 @@ namespace GoogleMapsComponents.Maps.Extension
             {
                 return ComputeEmptyResult<string>();
             }
-        }
-
-        public Task<Dictionary<string, bool>> GetDraggables(List<string> filterKeys = null)
-        {
-            List<string> matchingKeys = ComputeMathingKeys(filterKeys);
-
-            if (matchingKeys.Any())
-            {
-                Dictionary<Guid, string> internalMapping = ComputeInternalMapping(matchingKeys);
-                Dictionary<Guid, object> dictArgs = ComputeDictArgs(matchingKeys);
-
-                return _jsObjectRef.InvokeMultipleAsync<bool>(
-                    "getDraggable",
-                    dictArgs).ContinueWith(e => e.Result.ToDictionary(r => internalMapping[new Guid(r.Key)], r => r.Value));
-            }
-            else
-            {
-                return ComputeEmptyResult<bool>();
-            }
-        }
+        }        
 
         public Task<Dictionary<string, OneOf<string, Icon, Symbol>>> GetIcons(List<string> filterKeys = null)
         {
@@ -287,26 +162,7 @@ namespace GoogleMapsComponents.Maps.Extension
             {
                 return ComputeEmptyResult<string>();
             }
-        }
-
-        public Task<Dictionary<string, Map>> GetMaps(List<string> filterKeys = null)
-        {
-            List<string> matchingKeys = ComputeMathingKeys(filterKeys);
-
-            if (matchingKeys.Any())
-            {
-                Dictionary<Guid, string> internalMapping = ComputeInternalMapping(matchingKeys);
-                Dictionary<Guid, object> dictArgs = ComputeDictArgs(matchingKeys);
-
-                return _jsObjectRef.InvokeMultipleAsync<Map>(
-                    "getMap",
-                    dictArgs).ContinueWith(e => e.Result.ToDictionary(r => internalMapping[new Guid(r.Key)], r => r.Value));
-            }
-            else
-            {
-                return ComputeEmptyResult<Map>();
-            }
-        }
+        }        
 
         public Task<Dictionary<string, LatLngLiteral>> GetPositions(List<string> filterKeys = null)
         {
@@ -363,26 +219,7 @@ namespace GoogleMapsComponents.Maps.Extension
             {
                 return ComputeEmptyResult<string>();
             }
-        }
-
-        public Task<Dictionary<string, bool>> GetVisibles(List<string> filterKeys = null)
-        {
-            List<string> matchingKeys = ComputeMathingKeys(filterKeys);
-
-            if (matchingKeys.Any())
-            {
-                Dictionary<Guid, string> internalMapping = ComputeInternalMapping(matchingKeys);
-                Dictionary<Guid, object> dictArgs = ComputeDictArgs(matchingKeys);
-
-                return _jsObjectRef.InvokeMultipleAsync<bool>(
-                    "getVisible",
-                    dictArgs).ContinueWith(e => e.Result.ToDictionary(r => internalMapping[new Guid(r.Key)], r => r.Value));
-            }
-            else
-            {
-                return ComputeEmptyResult<bool>();
-            }
-        }
+        }        
 
         public Task<Dictionary<string, int>> GetZIndexes(List<string> filterKeys = null)
         {
@@ -432,15 +269,7 @@ namespace GoogleMapsComponents.Maps.Extension
             return _jsObjectRef.InvokeMultipleAsync(
                 "setCursor",
                 dictArgs);
-        }
-
-        public Task SetDraggables(Dictionary<string, bool> flags)
-        {
-            Dictionary<Guid, object> dictArgs = flags.ToDictionary(e => Markers[e.Key].Guid, e => (object)e.Value);
-            return _jsObjectRef.InvokeMultipleAsync(
-                "setDraggable",
-                dictArgs);
-        }
+        }        
 
         public Task SetIcons(Dictionary<string, string> icons)
         {
@@ -464,20 +293,7 @@ namespace GoogleMapsComponents.Maps.Extension
             return _jsObjectRef.InvokeMultipleAsync(
                 "setLabel",
                 dictArgs);
-        }
-
-        /// <summary>
-        /// Renders the marker on the specified map or panorama. 
-        /// If map is set to null, the marker will be removed.
-        /// </summary>
-        /// <param name="map"></param>
-        public async Task SetMaps(Dictionary<string, Map> maps)
-        {
-            Dictionary<Guid, object> dictArgs = maps.ToDictionary(e => Markers[e.Key].Guid, e => (object)e.Value);
-            await _jsObjectRef.InvokeMultipleAsync(
-                   "setMap",
-                   dictArgs);
-        }
+        }        
 
         public Task SetOpacities(Dictionary<string, float> opacities)
         {
@@ -485,15 +301,7 @@ namespace GoogleMapsComponents.Maps.Extension
             return _jsObjectRef.InvokeMultipleAsync(
                 "setOpacity",
                 dictArgs);
-        }
-
-        public Task SetOptions(Dictionary<string, MarkerOptions> options)
-        {
-            Dictionary<Guid, object> dictArgs = options.ToDictionary(e => Markers[e.Key].Guid, e => (object)e.Value);
-            return _jsObjectRef.InvokeMultipleAsync(
-                "setOptions",
-                dictArgs);
-        }
+        }        
 
         public Task SetPositions(Dictionary<string, LatLngLiteral> latLngs)
         {
@@ -517,15 +325,7 @@ namespace GoogleMapsComponents.Maps.Extension
             return _jsObjectRef.InvokeMultipleAsync(
                 "setTitle",
                 dictArgs);
-        }
-
-        public Task SetVisibles(Dictionary<string, bool> visibles)
-        {
-            Dictionary<Guid, object> dictArgs = visibles.ToDictionary(e => Markers[e.Key].Guid, e => (object)e.Value);
-            return _jsObjectRef.InvokeMultipleAsync(
-                "setVisible",
-                dictArgs);
-        }
+        }        
 
         public Task SetZIndexes(Dictionary<string, int> zIndexes)
         {
