+++ GoogleMapsComponents/Maps/MarkerOptions.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace GoogleMapsComponents.Maps
 {
-    public class MarkerOptions
+    public class MarkerOptions: BaseListableEntityOptions
     {
         /// <summary>
         /// The offset from the marker's position to the tip of an InfoWindow that has been opened with the marker as anchor.
@@ -17,13 +17,7 @@ namespace GoogleMapsComponents.Maps
         /// <summary>
         /// Which animation to play when marker is added to a map.
         /// </summary>
-        public Animation? Animation { get; set; }
-
-        /// <summary>
-        /// If true, the marker receives mouse and touch events. 
-        /// Default value is true.
-        /// </summary>
-        public bool? Clickable { get; set; }
+        public Animation? Animation { get; set; }       
 
         /// <summary>
         /// If false, disables cross that appears beneath the marker when dragging. 
@@ -34,12 +28,7 @@ namespace GoogleMapsComponents.Maps
         /// <summary>
         /// Mouse cursor to show on hover
         /// </summary>
-        public string Cursor { get; set; }
-
-        /// <summary>
-        /// If true, the marker can be dragged. Default value is false.
-        /// </summary>
-        public bool? Draggable { get; set; }
+        public string Cursor { get; set; }        
 
         /// <summary>
         /// Icon for the foreground. 
@@ -52,13 +41,7 @@ namespace GoogleMapsComponents.Maps
         /// Adds a label to the marker. The label can either be a string, or a MarkerLabel object.
         /// </summary>
         [JsonConverter(typeof(OneOfConverter))]
-        public OneOf<string, MarkerLabel> Label { get; set; }
-
-        /// <summary>
-        /// Map on which to display Marker.
-        /// </summary>
-        [JsonConverter(typeof(JsObjectRefConverter<Map>))]
-        public Map Map { get; set; }
+        public OneOf<string, MarkerLabel> Label { get; set; }        
 
         /// <summary>
         /// The marker's opacity between 0.0 and 1.0.
@@ -85,17 +68,6 @@ namespace GoogleMapsComponents.Maps
         /// <summary>
         /// Rollover text
         /// </summary>
-        public string Title { get; set; }
-
-        /// <summary>
-        /// If true, the marker is visible
-        /// </summary>
-        public bool? Visible { get; set; }
-
-        /// <summary>
-        /// All markers are displayed on the map in order of their zIndex, with higher values displaying in front of markers with lower values. 
-        /// By default, markers are displayed according to their vertical position on screen, with lower markers appearing in front of markers further up the screen.
-        /// </summary>
-        public int? ZIndex { get; set; }
+        public string Title { get; set; }        
     }
 }
